requirements.txt
langchain_community==0.3.19
langchain_experimental==0.3.4
langchain_ollama==0.2.3
pypdf==5.4.0
chromadb==1.0.3
six==1.17.0
